Creating `ErrorBoundary` component, consolidating error catching.
